requirements.txt
keras
matplotlib
numpy
opencv_python
Pillow
tensorflow
ttkbootstrap
